% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_goal.R
\name{goal}
\alias{goal}
\title{Methods to achieve a target sensitivity}
\usage{
goal(x, ...)
}
\arguments{
\item{...}{Options dispatched internally.}
}
\value{
Returns a \code{goal} object.
}
\description{
Methods to achieve a target sensitivity
}
