% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_influence.R
\name{infl}
\alias{infl}
\title{Methods to compute influence for `lm`, `ivreg`, and `matrix` objects}
\usage{
infl(x, ...)
}
\arguments{
\item{...}{Options dispatched internally.}
}
\value{
Returns a \code{influence} object.
}
\description{
Methods to compute influence for `lm`, `ivreg`, and `matrix` objects
}
