% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_init.R
\name{init}
\alias{init}
\title{Methods to compute an initial approximation of influence}
\usage{
init(x, ...)
}
\arguments{
\item{...}{Options dispatched internally.}
}
\value{
Returns a \code{initial} object.
}
\description{
Methods to compute an initial approximation of influence
}
