% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_plot.R
\name{plot.influence}
\alias{plot.influence}
\title{Boxplots for influence objects}
\usage{
\method{plot}{influence}(x, type = c("beta_i", "se_i", "tstat_i"), position, ...)
}
\arguments{
\item{x}{A \code{influence} object, obtained from \code{\link{infl}}.}

\item{type}{Optional character vector with the type of measure to plot.}

\item{position}{Numeric vector. Position of the variables.}

\item{...}{Not used.}
}
\value{
Returns \emph{x} invisibly.
}
\description{
Boxplots for influence objects
}
