% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_plot.R
\name{plot.sensitivity}
\alias{plot.sensitivity}
\title{Plotting method for sensitivity objects}
\usage{
\method{plot}{sensitivity}(x, type = c("path", "masking"), ...)
}
\arguments{
\item{x}{A \code{sensitivity} object, obtained from \code{\link{sens}}.}

\item{type}{Optional character vector with the type of plot desired.}

\item{...}{Dispatched internally. Argument \code{n} to limit the number of
plotted removals, and \code{threshold} to mark influence thresholds.}
}
\value{
Returns \emph{x} invisibly.
}
\description{
Plotting method for sensitivity objects
}
