% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_sens.R
\name{sens}
\alias{sens}
\title{Methods to compute sensitivity with `lm`, `ivreg`, and `influence` objects}
\usage{
sens(x, ...)
}
\arguments{
\item{...}{Options dispatched internally.}
}
\value{
Returns a \code{sensitivitiy} object.
}
\description{
Methods to compute sensitivity with `lm`, `ivreg`, and `influence` objects
}
