% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_options.R
\name{set_compute}
\alias{set_compute}
\title{Options for influence computation}
\usage{
set_compute(
  x = c("some", "none", "all"),
  hat,
  beta,
  sigma,
  se,
  tstat,
  cooksd,
  dffits,
  rstudent,
  covratio,
  cluster
)
}
\arguments{
\item{x}{Character scalar with short-hand options.}

\item{hat, beta, sigma, se, tstat, cooksd, dffits, rstudent, covration, cluster}{Logical scalars determining which quantities to compute (exactly).}
}
\value{
Returns a list with options.
}
\description{
Options for influence computation
}
