% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_lambda.R
\name{set_lambda}
\alias{set_lambda}
\title{Options for the influence type}
\usage{
set_lambda(
  type = c("beta_i", "sigma_i", "se_i", "tstat_i", "cooksd", "dffits", "rstudent",
    "covratio", "BKW"),
  position = 1L,
  sign = 1L,
  f = NULL
)
}
\arguments{
\item{type}{Character scalar with short-hand influence types.}

\item{position}{Integer scalar with the position of the variable of interest.}

\item{sign}{Integer scalar with the sign to apply to the influence.}

\item{f}{Optional custom function to calculate influence manually.}
}
\value{
Returns a function to compute lambda, with information in attributes.
}
\description{
Options for the influence type
}
