% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_options.R
\name{set_options}
\alias{set_options}
\title{Options for influence calculation}
\usage{
set_options(
  ...,
  p_max = NULL,
  n_max = NULL,
  sm_re = 1L,
  fwl = 0L,
  fwl_re = 1L,
  adaptive = TRUE
)
}
\arguments{
\item{...}{Dispatched internally to \code{\link{set_compute}}.}

\item{p_max, n_max}{Numeric scalar with the maximum percentage / number of
removals to consider.}

\item{sm_re}{Integer scalar with the number of steps until inverses are
recalculated exactly (instead of using the Sherman-Morrison update).}

\item{fwl}{Optional integer scalar with the position of variables to
marginalise out.}

\item{fwl_re}{Optional integer scalar with the steps until variables are
marginalised out again.}

\item{adaptive}{Logical scalar setting adaptive set identificaiton.}
}
\value{
Returns a list with options.
}
\description{
Options for influence calculation
}
