% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_lambda.R
\name{set_target}
\alias{set_target}
\title{Options to set a target influence}
\usage{
set_target(
  target = 0,
  type = c("less", "leq", "geq", "greater"),
  f = function(x, y, ...) {
     NULL
 }
)
}
\arguments{
\item{target}{Numerical vector with the target value(s).}

\item{type}{Character scalar with the tpye of inequality.}

\item{f}{Optional custom function to compare values.}
}
\value{
Returns a function to check whether a target value is reached.
}
\description{
Options to set a target influence
}
