% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/81_summary.R
\name{summary.sensitivity}
\alias{summary.sensitivity}
\title{Summary method for sensitivity objects}
\usage{
\method{summary}{sensitivity}(x, n = 0, threshold = qnorm(0.975), ...)
}
\arguments{
\item{x}{A \code{sensitivity} object, obtained from \code{\link{sens}}.}

\item{n}{Optional character vector with a maximum of removals.}

\item{threshold}{Optional numerical vector with influence thresholds.}

\item{...}{Not used.}
}
\value{
Returns a list with summary results.
}
\description{
Summary method for sensitivity objects
}
